- refactoring FsService, moved types definition outside main service for easier inclusion from other files ( ex: worker )
