+++ src/app/services/fs-service/fs.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { IndexeddbFsDriver } from './fs.service.test';
+import { FsNodeFolder, FsNodeFile, FsServiceDriver } from './fs.service.types';
 
-export class FsNodeList extends Array<FsNodeFile|FsNodeFolder>{};
 
 @Injectable({
   providedIn: 'root'
@@ -190,47 +190,6 @@ export class Tar{
   
   
 
-}
-
-export interface FsNode {
-  name: string;
-  path: string;
-}
-
-export interface FsNodeFolder extends FsNode {
-  folders: FsNodeFolder[];
-  files: FsNodeFile[];
-}
-export interface FsNodeFile extends FsNode {
-  content:string|ArrayBuffer
-}
-
-
-
-
-
-export interface FsServiceDriver {
-  rootDir:string;
-  
-  mount(path: string): Promise<boolean>;
-  
-  unmount(path: string): Promise<boolean>;
-
-  ready(): Promise<boolean>;
-
-  createDirectory(fullpath:string): Promise<boolean>;
-
-  writeFile(fullpath:string, content:string|ArrayBuffer): Promise<number>;
-
-  readFile(fullpath:string, binary?: boolean): Promise<string|ArrayBuffer|null>;
-
-  readDirectory(fullpath:string): Promise<FsNodeFolder|null>;
-
-  scanDirectory(fullpath:string): Promise<FsNodeFolder|null>;
-
-  delete(fullpath:string): Promise<boolean>;
-
-  exists(fullpath:string): Promise<boolean>;
 }
